+++ Flowery.NET.Gallery.Android/MainActivity.cs
@@ -1,4 +1,4 @@
-﻿using Android.App;
+using Android.App;
 using Android.Content.PM;
 using Avalonia;
 using Avalonia.Android;
@@ -15,6 +15,7 @@ public class MainActivity : AvaloniaMainActivity<App>
 {
     protected override AppBuilder CustomizeAppBuilder(AppBuilder builder)
     {
-        return base.CustomizeAppBuilder(builder);
+        return base.CustomizeAppBuilder(builder)
+            .WithInterFont();
     }
 }