+++ Flowery.NET.Gallery.Desktop/Program.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using Avalonia;
 
 
@@ -17,5 +17,6 @@ sealed class Program
     public static AppBuilder BuildAvaloniaApp()
         => AppBuilder.Configure<App>()
             .UsePlatformDetect()
+            .WithInterFont()
             .LogToTrace();
 }